Adding an IntelliJ code style shared configuration file for Rave.  This actually doesnt have any changes from the IntelliJ default right now but I'm guessing in the future we'll need to make at least some changes so this should let us share those changes.
On my workstation this file lives at:

$home/.IntelliJIdea10/config/codestyles/RaveCodeFormatter.xml

If you copy it there and restart IntelliJ it should pick it up and make it an available as an option to use for the project.  Anytime we change it we'll have to copy it from there and commit it back to source control.  IntelliJ doesnt seem to have the option to store it/read it from an alternate location.  References here:

http://www.jetbrains.com/idea/webhelp/exporting-project-code-style-settings.html

http://musingsofaprogrammingaddict.blogspot.com/2010/03/import-code-style-settings-into.html

git-svn-id: https://svn.apache.org/repos/asf/incubator/rave/trunk@1092698 13f79535-47bb-0310-9956-ffa450edef68