RAVE-63: maven-remote-resources-plugin (m-r-r-p) will automatically merge/append matching resources when put under src/main/appended-resources For the LICENSE and NOTICE files, which are automatically/default added as configured through the Apache master pom, we therefore only have to specify required additional LICENSE and NOTICE snippets within there.
As example I'm adding a first required NOTICE addition for the H2 database bundled with rave-shindig.
And, as we regrettably also manually need to maintain the (complete) LICENSE and NOTICE files for the assembled distributions, e.g. convering *all* bundled usages,
I've also added the H2 license notification to the dist/NOTICE file.

There is still plenty more required attribution to do, this is just a first next step... 

git-svn-id: https://svn.apache.org/repos/asf/incubator/rave/trunk@1139242 13f79535-47bb-0310-9956-ffa450edef68